Move GetStartOnSystemStartup and SetStartOnSystemStartup to GUI code
Reference: https://github.com/bitcoin/bitcoin/pull/1178
