+++ src/qt/bitcoin.cpp
@@ -289,6 +289,10 @@ int main(int argc, char *argv[])
 
     try
     {
+        // Regenerate startup link, to fix links to old versions
+        if (GUIUtil::GetStartOnSystemStartup())
+            GUIUtil::SetStartOnSystemStartup(true);
+
         BitcoinGUI window;
         guiref = &window;
         // Set in AppInit for daemon so that RPC works properly
